@@ -1,13 +1,13 @@
 module github.com/vespa-engine/vespa/client/go
 
-go 1.20
+go 1.22.4
 
 require (
 	github.com/alessio/shellescape v1.4.2
 	github.com/briandowns/spinner v1.23.1
 	github.com/fatih/color v1.17.0
-	// This is the most recent version compatible with Go 1.20. Upgrade when we upgrade our Go version
-	github.com/go-json-experiment/json v0.0.0-20230324203220-04923b7a9528
+	// This is the most recent version compatible with Go 1.21. Upgrade when we upgrade our Go version
+	github.com/go-json-experiment/json v0.0.0-20240412061110-8868a69194fa
 	github.com/klauspost/compress v1.17.9
 	github.com/mattn/go-colorable v0.1.13
 	github.com/mattn/go-isatty v0.0.20
