@@ -478,10 +478,7 @@ public class SelectParser implements Parser {
 
         if (annotations != null) {
             annotations.traverse((ObjectTraverser) (annotation_name, annotation_value) -> {
-                if (TARGET_HITS.equals(annotation_name)){
-                    weakAnd.setN((int)(annotation_value.asDouble()));
-                }
-                if (TARGET_NUM_HITS.equals(annotation_name)) {
+                if (TARGET_HITS.equals(annotation_name) || TARGET_NUM_HITS.equals(annotation_name)){
                     weakAnd.setN((int)(annotation_value.asDouble()));
                 }
             });
