@@ -2,8 +2,6 @@
 
 package com.yahoo.vespasignificance;
 
-import com.fasterxml.jackson.annotation.JsonAutoDetect;
-import com.fasterxml.jackson.annotation.PropertyAccessor;
 import com.fasterxml.jackson.core.JsonFactory;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.ObjectWriter;
@@ -27,6 +25,8 @@ import com.yahoo.language.process.Tokenizer;
 import com.yahoo.language.significance.impl.DocumentFrequencyFile;
 import com.yahoo.language.significance.impl.SignificanceModelFile;
 import com.yahoo.text.Utf8;
+import io.airlift.compress.zstd.ZstdInputStream;
+import io.airlift.compress.zstd.ZstdOutputStream;
 
 import java.io.IOException;
 import java.io.BufferedReader;
@@ -34,6 +34,9 @@ import java.io.ByteArrayInputStream;
 import java.io.File;
 import java.io.InputStream;
 import java.io.InputStreamReader;
+import java.io.FileOutputStream;
+import java.io.FileInputStream;
+import java.io.OutputStream;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
@@ -59,6 +62,7 @@ public class SignificanceModelGenerator {
 
     final DocumentTypeManager types = new DocumentTypeManager();
     final DocumentType docType;
+    private final boolean useZstCompression;
     private final static String VERSION = "1.0";
     private final static String ID = "1";
     private final static String SIGNIFICANCE_DESCRIPTION = "Significance model for input file";
@@ -66,14 +70,24 @@ public class SignificanceModelGenerator {
 
     public SignificanceModelGenerator(ClientParameters clientParameters) {
         this.clientParameters = clientParameters;
+
+        if (clientParameters.zstCompression && !clientParameters.outputFile.endsWith(".zst")) {
+            throw new IllegalArgumentException("Output file must have .zst extension when using zst compression");
+        }
+
+        language = Language.fromLanguageTag(clientParameters.language);
+        if (language == Language.UNKNOWN) {
+            throw new IllegalArgumentException("Unknown language: " + clientParameters.language);
+        }
+
         OpenNlpLinguistics openNlpLinguistics = new OpenNlpLinguistics();
         tokenizer = openNlpLinguistics.getTokenizer();
         objectMapper = new ObjectMapper();
 
-        language = Language.fromLanguageTag(clientParameters.language);
-
         docType = new DocumentType(clientParameters.docType);
         docType.addField(new Field(clientParameters.field, DataType.STRING));
+        useZstCompression = clientParameters.zstCompression;
+
         types.registerDocumentType(docType);
     }
 
@@ -103,8 +117,14 @@ public class SignificanceModelGenerator {
         long pageCount = i - 1;
 
         SignificanceModelFile modelFile;
-        if (Paths.get(clientParameters.outputFile).toFile().exists()) {
-            modelFile = objectMapper.readValue(new File(clientParameters.outputFile), SignificanceModelFile.class);
+        File outputFile = Paths.get(clientParameters.outputFile).toFile();
+        if (outputFile.exists()) {
+
+            InputStream in = outputFile.toString().endsWith(".zst") ?
+                    new ZstdInputStream(new FileInputStream(outputFile)) :
+                    new FileInputStream(outputFile);
+
+            modelFile = objectMapper.readValue(in, SignificanceModelFile.class);
 
             modelFile.addLanguage(clientParameters.language, new DocumentFrequencyFile(DOC_FREQ_DESCRIPTION, pageCount, getFinalDocumentFrequency()));
 
@@ -117,7 +137,12 @@ public class SignificanceModelGenerator {
         }
         try {
             ObjectWriter writer = objectMapper.writerWithDefaultPrettyPrinter();
-            writer.writeValue(new File(clientParameters.outputFile), modelFile);
+
+            OutputStream out = useZstCompression ?
+                    new ZstdOutputStream(new FileOutputStream(clientParameters.outputFile)) :
+                    new FileOutputStream(clientParameters.outputFile);
+
+            writer.writeValue(out, modelFile);
         } catch (IOException e) {
             throw new IllegalStateException("Failed to write model to output file", e);
         }
