@@ -1,8 +1,11 @@
 // Copyright Vespa.ai. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
 
 #include "test_hook.h"
+#include <vespa/vespalib/util/count_down_latch.h>
+#include <vespa/vespalib/util/barrier.h>
+#include <vespa/vespalib/util/thread.h>
 #include <vespa/vespalib/util/stringfmt.h>
-#include <vespa/vespalib/util/size_literals.h>
+#include <cassert>
 #include <regex>
 
 namespace vespalib {
@@ -29,7 +32,7 @@ TestThreadWrapper::threadEntry()
         master.check(false, __FILE__, __LINE__, "test threw stuff", false);
     }
     _barrier.destroy();
-    master.setThreadBarrier(0);
+    master.setThreadBarrier(nullptr);
     bool fail = (master.getThreadFailCnt() > preThreadFailCnt);
     master.setThreadUnwind(false);
     master.setThreadIgnore(false);
@@ -39,30 +42,64 @@ TestThreadWrapper::threadEntry()
     master.setThreadName(oldThreadName.c_str());
 }
 
-TestHook *TestHook::_head = 0;
-TestHook *TestHook::_tail = 0;
+TestHook *TestHook::_head = nullptr;
+TestHook *TestHook::_tail = nullptr;
+
+TestHook::~TestHook() = default;
 
 TestHook::TestHook(const std::string &file, const std::string &name, bool ignore)
-    : _next(0),
+    : _next(nullptr),
       _name(name),
       _tag(make_string("%s:%s", file.c_str(), name.c_str())),
       _ignore(ignore)
 {
-    if (_head == 0) {
-        assert(_tail == 0);
+    if (_head == nullptr) {
+        assert(_tail == nullptr);
         _head = this;
         _tail = this;
     } else {
-        assert(_tail != 0);
-        assert(_tail->_next == 0);
+        assert(_tail != nullptr);
+        assert(_tail->_next == nullptr);
         _tail->_next = this;
         _tail = this;
     }
 }
 
+bool TestHook::runMyTest(const FixtureFactory & fixture_factory, size_t num_threads) {
+    assert(num_threads > 0);
+    using ThreadUP = std::unique_ptr<TestThreadWrapper>;
+    using FixtureUP = std::unique_ptr<TestFixtureWrapper>;
+    std::vector<TestMaster::TraceItem> traceStack = TestMaster::master.getThreadTraceStack();
+    CountDownLatch latch(num_threads);
+    Barrier barrier(num_threads);
+    std::vector<FixtureUP> fixtures;
+    std::vector<ThreadUP> threads;
+    ThreadPool pool;
+    threads.reserve(num_threads);
+    fixtures.reserve(num_threads);
+    for (size_t i = 0; i < num_threads; ++i) {
+        FixtureUP fixture_up = fixture_factory();
+        fixture_up->thread_id = i;
+        fixture_up->num_threads = num_threads;
+        threads.emplace_back(new TestThreadWrapper(_ignore, latch, barrier, traceStack, *fixture_up));
+        fixtures.push_back(std::move(fixture_up));
+    }
+    for (size_t i = 1; i < num_threads; ++i) {
+        pool.start([&target = *threads[i]](){ target.threadEntry(); });
+    }
+    threads[0]->threadEntry();
+    latch.await();
+    pool.join();
+    bool result = true;
+    for (size_t i = 0; i < num_threads; ++i) {
+        result = result && threads[i]->getResult();
+    }
+    return result;
+}
+
 const char *lookup_subset_pattern(const std::string &name) {
     const char *pattern = getenv("TEST_SUBSET");
-    if (pattern != 0) {
+    if (pattern != nullptr) {
         fprintf(stderr, "%s: info:  only running tests matching '%s'\n",
                 name.c_str(), pattern);
     } else {
@@ -80,7 +117,7 @@ TestHook::runAll()
     size_t testsFailed = 0;
     size_t testsIgnored = 0;
     size_t testsSkipped = 0;
-    for (TestHook *test = _head; test != 0; test = test->_next) {
+    for (TestHook *test = _head; test != nullptr; test = test->_next) {
         if (std::regex_search(test->_tag, pattern)) {
             bool ignored = test->_ignore;
             bool failed = !test->run();
