@@ -2,6 +2,7 @@
 
 #include <vespa/vespalib/testkit/test_kit.h>
 #include <unistd.h>
+#include <cassert>
 
 using namespace vespalib;
 
@@ -9,9 +10,9 @@ void * hammer(void * arg)
 {
     usleep(4000000);
     long seconds = * static_cast<const long *>(arg);
-    long stopTime(time(NULL) + seconds);
+    long stopTime(time(nullptr) + seconds);
     pthread_t id = pthread_self();
-    while (time(NULL) < stopTime) {
+    while (time(nullptr) < stopTime) {
          std::vector<pthread_t *> allocations;
          for (size_t i(0); i < 2000; i++) {
              pthread_t *t = new pthread_t[20];
@@ -21,11 +22,11 @@ void * hammer(void * arg)
              }
          }
 
-         for (size_t i(0); i < allocations.size(); i++) {
+         for (auto & allocation : allocations) {
              for (size_t j(0); j < 20; j++) {
-                 assert(allocations[i][j] == id);
+                 assert(allocation[j] == id);
              }
-             delete [] allocations[i];
+             delete [] allocation;
          }
     }
     return arg;
@@ -35,9 +36,9 @@ TEST_MAIN() {
     size_t threadCount(1024);
     long seconds(10);
     if (argc >= 2) {
-        threadCount = strtoul(argv[1], NULL, 0);
+        threadCount = strtoul(argv[1], nullptr, 0);
         if (argc >= 3) {
-            seconds = strtoul(argv[2], NULL, 0);
+            seconds = strtoul(argv[2], nullptr, 0);
         }
     }
 
